MOS-28568 Fix in salt based hashing and encryption.

Former-commit-id: 46b392bdc8c4b6e4a8b8409e4d95aee607121219
Former-commit-id: eb430a8f584f124e899428f7b77bd827b476d7c6

+++ authentication/authentication-common/src/main/java/io/mosip/authentication/common/service/builder/AuthTransactionBuilder.java
@@ -203,10 +203,10 @@ public class AuthTransactionBuilder {
 				// Setting primary code only
 				autnTxn.setLangCode(env.getProperty(IdAuthConfigKeyConstants.MOSIP_PRIMARY_LANGUAGE));
 				int saltModuloConstant = env.getProperty(IdAuthConfigKeyConstants.UIN_SALT_MODULO, Integer.class);
-				int uinModulo = (int) (Long.parseLong(uin) % saltModuloConstant);
+				Long uinModulo = (Long.parseLong(uin) % saltModuloConstant);
 				String hashSaltValue = uinHashSaltRepo.retrieveSaltById(uinModulo);
 				autnTxn.setUinHash(HMACUtils.digestAsPlainTextWithSalt(uin.getBytes(), hashSaltValue.getBytes()));
-				String encryptSaltValue = uinEncryptSaltRepo.retrieveSaltById(uinModulo);
+				String encryptSaltValue = uinEncryptSaltRepo.retrieveSaltById(uinModulo.intValue());
 				autnTxn.setUin(uinModulo + IdAuthCommonConstants.UIN_MODULO_SPLITTER + uin
 						+ IdAuthCommonConstants.UIN_MODULO_SPLITTER + encryptSaltValue);
 				Optional<String> clientId = Optional.of(transactionManager.getUser());
